use bench library for bench project
